expansion to address issue #6
Substantial changes to modify behaviour in combining data from the retrieved product with the input data. Dual behaviour now stacks the data if product is of class move2_loc, otherwise attaches the data to the appended_products` attribute, as before. Changes include updates to testing, documentation and App user settings.